Add files via upload
To run it upload the code inside the arduino uno bot file to arduino ide and upload the code to arduino uno with the ide then open the arduino ide terminal and write the questions wrote inside the arduino uno bot code to have a reply, It can be expanded with more ram size like an arduino due. with somesensors ad reprogramming the code you cna adda mic a videocamera etcs to build a little or bigger robot.